ciao-scheduler: make ssntp testutil test diff-able
I'm splitting the ssntp test out into its own file scheduler_ssntp_test.go
which can then be easily diff'd against testutil/clien_server_test.go
to show how the fully synthetic test was modified to use the scheduler
server instead of the testutil server.  That will ease maintenance and
provide a better complex case example usage of testutil.
